LPS-104435 Move logic from GradleVersionCheck to GradleDependencyArtifactsCheck
