+++ source-formatter.properties
@@ -233,8 +233,6 @@
 
     source.check.gradle.dependency.artifacts.check.renameArtifacts=
 
-    source.check.gradle.version.check.enabled=true
-
     source.check.java.component.activate.check.enabled=true
 
     source.check.java.module.component.check.allowedClassNames=\
